pyproject.toml
[project]
name = "comfyui-starwrapper"
description = "A ComfyUI custom node wrapper for [STAR (Spatial-Temporal Augmentation with Text-to-Video Models for Real-World Video Super-Resolution)](https://github.com/NJU-PCALab/STAR), enabling high-quality video upscaling with AI-powered super-resolution."
version = "0.1.4"
license = {file = "LICENSE"} 
classifiers = [
    # For OS-independent nodes (works on all operating systems)
    "Operating System :: OS Independent",
    # GPU Accelerator support. Pick the ones that are supported by your extension.
    "Environment :: GPU :: NVIDIA CUDA",    # NVIDIA CUDA support
]

dependencies = ["accelerate", "av", "opencv-python", "easydict", "einops", "open-clip-torch", "fairscale", "torchsde", "pytorch-lightning", "diffusers[torch]", "huggingface_hub", "numpy<2", "ruff", "transformers", "bitsandbytes"]

[project.urls]
Repository = "https://github.com/vjumpkung/ComfyUI-STARWrapper"
#  Used by Comfy Registry https://registry.comfy.org
Documentation = "https://github.com/vjumpkung/ComfyUI-STARWrapper/wiki"
"Bug Tracker" = "https://github.com/vjumpkung/ComfyUI-STARWrapper/issues"

[tool.comfy]
PublisherId = "vjumpkung"
DisplayName = "ComfyUI-STARWrapper"
Icon = ""
includes = [] 
# "requires-comfyui" = ">=1.0.0"  # ComfyUI version compatibility

